refactor: confirm_diallog.dart를 사용하도록 가져오기 경로를 업데이트하고 명확성을 위해 파일 구조를 재구성
