+++ lib/page/myPage/mypage.dart
@@ -2,8 +2,8 @@ import 'package:contained_tab_bar_view/contained_tab_bar_view.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:lockerroom/const/color.dart';
-import 'package:lockerroom/my_post/my_market_page.dart';
-import 'package:lockerroom/my_post/mypost.dart';
+import 'package:lockerroom/page/my_post/my_market_page.dart';
+import 'package:lockerroom/page/my_post/mypost.dart';
 import 'package:lockerroom/page/alert/confirm_diallog.dart';
 import 'package:lockerroom/page/follow/follow_list_page.dart';
 import 'package:lockerroom/page/intution_record/intution_record_list_page.dart';
