+++ lib/page/myPage/user_detail_page.dart
@@ -5,8 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:lockerroom/const/color.dart';
 import 'package:lockerroom/model/post_model.dart';
 import 'package:lockerroom/model/team_model.dart';
-import 'package:lockerroom/my_post/user_market_post_page.dart';
-import 'package:lockerroom/my_post/user_post_page.dart';
+import 'package:lockerroom/page/my_post/user_market_post_page.dart';
+import 'package:lockerroom/page/my_post/user_post_page.dart';
 import 'package:lockerroom/page/follow/follow_list_page.dart';
 import 'package:lockerroom/provider/feed_provider.dart';
 import 'package:lockerroom/provider/follow_provider.dart';
