+++ lib/page/my_post/my_market_page.dart
@@ -1,52 +1,59 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:lockerroom/const/color.dart';
 import 'package:lockerroom/model/market_post_model.dart';
 import 'package:lockerroom/page/afterMarket/after_market_detail_page.dart';
-import 'package:lockerroom/page/alert/diallog.dart';
+import 'package:lockerroom/page/alert/confirm_diallog.dart';
 import 'package:lockerroom/provider/market_feed_provider.dart';
 import 'package:lockerroom/provider/team_provider.dart';
 import 'package:provider/provider.dart';
 import 'package:toastification/toastification.dart';
 
-class UserMarketPostPage extends StatefulWidget {
-  final String userId;
-  const UserMarketPostPage({super.key, required this.userId});
+class MyMarketPage extends StatelessWidget {
+  const MyMarketPage({super.key});
 
-  @override
-  State<UserMarketPostPage> createState() => _UserMarketPostPageState();
-}
-
-class _UserMarketPostPageState extends State<UserMarketPostPage> {
   @override
   Widget build(BuildContext context) {
-    final tp = context.read<TeamProvider>();
+    final marketFeedProvider = Provider.of<MarketFeedProvider>(
+      context,
+      listen: false,
+    );
+    final tp = Provider.of<TeamProvider>(context, listen: false);
     final teamColor = tp.selectedTeam?.color;
-    final marketFeedProvider = context.read<MarketFeedProvider>();
     return Scaffold(
       backgroundColor: BACKGROUND_COLOR,
-      body: StreamBuilder<QuerySnapshot<Map<String, dynamic>>>(
-        stream: FirebaseFirestore.instance
-            .collection('market_posts')
-            .where('userId', isEqualTo: widget.userId)
-            .orderBy('createdAt', descending: true)
-            .snapshots(),
+      body: StreamBuilder<List<MarketPostModel>>(
+        stream: marketFeedProvider.listenMyMarketPosts(),
         builder: (context, snapshot) {
-          if (!snapshot.hasData)
+          if (snapshot.hasError) {
+            debugPrint('market 스트림에러 : ${snapshot.error}');
+            return Center(child: Text('에러 발생'));
+          }
+          if (!snapshot.hasData) {
             return Center(child: CircularProgressIndicator(color: teamColor));
-          final marketPosts = snapshot.data!.docs
-              .map((d) => MarketPostModel.fromDoc(d))
-              .toList();
+          }
 
-          if (marketPosts.isEmpty) return Center(child: Text('게시물이 없습니다'));
+          final posts = snapshot.data!;
+
+          if (posts.isEmpty) {
+            return Center(
+              child: Text(
+                '게시물이 없습니다',
+
+                style: TextStyle(
+                  fontSize: 16,
+                  color: GRAYSCALE_LABEL_500,
+                  fontWeight: FontWeight.bold,
+                ),
+              ),
+            );
+          }
 
           return ListView.builder(
-            itemCount: marketPosts.length,
-            itemBuilder: (_, i) {
-              final mp = marketPosts[i];
+            itemCount: posts.length,
+            itemBuilder: (context, index) {
               return MyMarketPostWidget(
-                marketPost: mp,
+                marketPost: posts[index],
                 marketFeed: marketFeedProvider,
               );
             },
